refactor(frontend): update inviteUser dialog
